[composebox] Make navigation follow omnibox flow in composebox.
This CL allows typed navigation to follow the omnibox autocomplete flow.
Currently, the ComposeboxHandler handles navigation through
`SubmitQuery()` whenever a user clicks submit. However this causes
omnibox metrics to not get logged for these navigations. To fix this,
run the SearchProvider and only return the default match for all typed
input. Then, call `openAutocompleteMatch` to follow the normal
autocomplete flow for opening a match. The navigation is still delegated
to ComposeboxHandler after all omnibox metrics are logged (in order to
correctly navigate to AIM).

One important note is that if only a file is uploaded and the user
clicks submit, since there is not a default match, these navigations
won't get recorded in Omnibox logs. This will follow the existing path.

NOTE: A follow up CL will do this for when the AIM button is clicked in
the realbox on the NTP.

Change-Id: I85111ba2adf8287cb3a885cf572d8d2ee8fdfa83
Bug: 439602823
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6849236
Commit-Queue: Nihar Majmudar <niharm@google.com>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Cr-Commit-Position: refs/heads/main@{#1506806}