[crd host][wayland] Set monitor scale without triggering the confirmation dialog
I have done some AI-guided code reading on the mutter code base and I've
found that the confirmation dialog is only shown when the
ApplyMonitorsConfig @method is persistent. Judging by the code, this is
most likely an intended behavior -- with temporary, the display config
is not written to the disk and will be gone after you reboot, and with
persistent it will stick. See comment in gnome_desktop_resizer.cc. For
CRD, there is no benefit using persistent, so this CL changes the method
to temporary and sets the monitor scale to the value proportionally
closest to the preferred scale.

Note that there are still two open issues/future works:

1. A mutter developer is working on the idea of setting scales and
   display layouts via PipeWire. See:
   https://gitlab.gnome.org/GNOME/mutter/-/issues/4275. If that ever
   gets implemented, then our code can be significantly simplified, and
   reliability can be improved since there won't be any more guessing
   for the connector name of a virtual stream. But for now, we still
   need to use the DisplayConfig API.
2. There seems to be a bug with setting fractional scales when there are
   multiple virtual monitors. See:
   https://gitlab.gnome.org/GNOME/mutter/-/issues/4277.
   For now, this CL disables fractional scales when there are multiple
   monitors.

Screencast to prove that this works:
http://screencast/cast/NjQ3NjQxMzE3MDM1MjEyOHw5ZGRmYzFmYi03NQ

Bug: 431816005
Change-Id: I75c1c16282fe7d766493ba491518eb86f4c47d79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6886455
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506835}