Change the return type of AndroidPrerenderManager.startPrerendering
This CL changes the return type of
AndroidPrerenderManager.startPrerendering from boolean to void.
The return value was used to do following things
1. Avoid calling `startPrerendering` if the url is the same as started
one.
2. Avoid calling `stopPrerendering` if there is no on-going prerender
to avoid cross-language(Java => CPP) communication overhead.
For (1), with the introduction of `NewTabPagePreloadPipelineManager`,
it is desired to avoid logic spreading in different layers to avoid
potential inconsistency in the future. It is also unknown that how
often the scenario happens as prerender is canceled on every onTouch
cancelation, two consecutive onTouch ACTION_DOWN doesn't look common.
For (2), `mPrerenderTriggered` is introduced to mitigate the overhead,
though there is behavioral change as the implementation no longer
checks the return value.

Bug: 421941586
Change-Id: Id12ce3feae883f05b3cb678aff7b098bf8a8c892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6865896
Commit-Queue: Huanpo Lin <robertlin@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506845}