flyout: Add hover listeners to ListMenu items
This CL adds hover listeners to items in `ListMenu` so that we can
create new flyout windows when items with submenus get hover.

We add the listeners to items that don't have submenus too, because
we will use the hover on those items to dismiss the flyout popup.

Bug: 433410990
Change-Id: I6f4885a149c65cfe87a365c7f557ac681e185e1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6874345
Auto-Submit: Masa Fujita <massan@google.com>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Commit-Queue: Masa Fujita <massan@google.com>
Cr-Commit-Position: refs/heads/main@{#1506883}