Update NewTabPagePreloadPipeline*::StartPrerender behavior
This CL makes following changes similar to http://crrev.com/c/6857734
1. Introduce `NewTabPagePreloadPipelineManager::EnsurePipelineForUrl`
which resets pipeline if url mismatches in `StartPrerender`. This
guarantees a new pipeline or a pipeline with the same url.
2. Remove return value in `NewTabPagePreloadPipeline::StartPrerender`
as the pipeline manager will not reset pipeline according to the
trigger result anymore.
3. Due to mechanism in 1, if a `prerender_handle_` exists, it must
have the same url. Remove `prerender_handle_->IsValid()` from the
duplicate check condition in
`NewTabPagePreloadPipeline::StartPrerender`.

Behavioral change:
1. Before the changes
`NewTabPagePreloadPipelineManager::StartPrerender` returns if
pipeline is not null, now it creates a new pipeline if the given url
is different.
2. `NewTabPagePreloadPipelineManager::StartPrerender` no longer
resets pipeline according to the result of
`NewTabPagePreloadPipeline::StartPrerender`.
We expect that 2 doesn't affect user's experience. 1 changes the behavior for
two mouseenters that don't contain mouseleave, but it's quite rare and safe.

Bug: 421941586
Change-Id: I845f4fffc303672397be2d7dafcb240a16b3d275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6865471
Commit-Queue: Huanpo Lin <robertlin@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506885}