Pass web visibility time window to `AssociateGoogleVariationID`.
This change introduces a `GetWebVisibilityTimeWindow` helper to extract
start and end dates from a `Study`. The `RegisterVariationIds` function
is updated to accept the full `Study` object, allowing it to compute and
pass this `TimeWindow` when calling `AssociateGoogleVariationID` for
both app and web experiment IDs.

This change updates the variations associated data with the timebox
information contained in the seed, allowing it to be applied as
variation headers and vectors are being calculated/recalculated over
time.

Bug: 422445047
Change-Id: I469ec7743e4727fbde6d7aadeee6505d2c04e4d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6846860
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506895}