passwords/desktop: Move "has autofillable pwds?" cache to memory
LoginDatabase::SetIsEmptyCb() was originally used to set the
kEmptyProfileStoreLoginDatabase pref on Android, which had to be checked
before PasswordStore instantiation. The method was later extended to set
the kAutofillableCredentials{Profile,Account}StoreLoginDatabase prefs
on desktop, which determine whether to offer the "Select password"
button in the context menu (the button should be shown if the user has
any autofillable passwords in either PasswordStore, regardless of the
origin).
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/autofill/autofill_context_menu_manager.cc;l=169-174;drc=5a67697c2b6446399af384d5cc57d5fe61f543c4

That last use can be reimplemented in-memory and the prefs removed.
This is a step towards removing SetIsEmptyCb(), which was quite
invasive to begin with. To achieve this we leverage the existing
PasswordCounter class, used only on iOS until now. The class is modified
to match the SetIsEmptyCb() implementation and exclude both federated
and username-only credentials, since the same product rationale
applies for both desktop and iOS.

Bug: None
Change-Id: I6e5efd7aa0e109bd5ef3b15db119b1d22a9a6c20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6818510
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1506916}