Enable new URLPattern canon by default and update tests to match
Update the tests related to URLPattern using dummy URL canonicalization
and enables the feature by default. The WPT -expected.txt files related
to URLPattern are removed so that we are now validating the standard
behavior.

Updated the webview_test.ts test that was broken by this change due to
it testing non-standard behavior of URLPattern. I've updated its
expectations to match standard behavior.

This CL is part of several CLs to address interop issues with
URLPattern. See bug https://crbug.com/409350827 for more information.

The bug fixes for URLPattern are edge case enough to not require a
feature flag, but it is still nice to have the feature flag to disable
the change if anything goes wrong. So, the feature flag is on by default
and exists as an emergency killswitch.

Enabled-by-default-reason: killswitch
Bug: 409350827
Change-Id: I9f59aaae9c56f469e6d934a3efbc9bd64dd03c65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6819984
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Takashi Nakayama <tnak@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Commit-Queue: Dave Risney <david.risney@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1506922}