Support CorrectEventAsNavigationOrActivationOrigined for non painting
This CL changes the implementation of
`CorrectEventAsNavigationOrActivationOrigined` to accept non-painting
events as well, which the event might happen before activation, but
might be recorded after actiavation (e.g. custom performance marks).

Instead of checking if the event is always greater than zero, we take
the max of zero and the adjusted time, so that if the event happens
before activation, we return zero.

Bug: 440247203, 436345871
Change-Id: I77cc3cb40d1621418503222d4ce0f11831940bb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6887671
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Keita Suzuki <suzukikeita@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1506938}