Announce password recovery suggestions for accessibility.
Adds a new method `MaybeAnnouncePasswordRecoveryPopup()` to `PopupViewViews`. This method checks if the first suggestion is a `kBackupPasswordEntry` and, if true, announces to screen reader users that password recovery suggestions are available. This announcement is triggered when the popup is shown and when suggestions are updated.

Change-Id: Ibd5f53087f236dca8e512c0f5422f77cf1208a23
Fixed: 438514797
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6874903
Reviewed-by: Bruno Braga <brunobraga@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Oleksandr Tara <otara@google.com>
Cr-Commit-Position: refs/heads/main@{#1506950}