Delete cronet_perf_test_apk
A follow-up CL is adding a dependency on //net/android:net_errors_java.
The existance of this targets complicates the dependency chain
noticeably.

This test has not been running in a long time (its bot has been deleted,
after it stayed broken for a long time). So this cleanup feels the right
thing to do.

Bug: b:439766861
Change-Id: Ifb77d05cc96a56bce16d388d12699ac03c3e5e8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6863029
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Stefano Duo <stefanoduo@google.com>
Reviewed-by: Mohannad Farrag <aymanm@google.com>
Cr-Commit-Position: refs/heads/main@{#1506992}