[ios] Fix crash caused in other test suites by improper cleanup
MagicStackRankingModelTest was installing an UkmClient for testing
but forgot to clear it. This caused an assertion failures when the
NewTabPageCoordinatorTest test suite later tried to access to the
UkmClient, as UkmDatabaseClient asserts that if a test client is
registered, then there must be one client per profile.

Update MagicStackRankingModelTest to clean up after itself in its
TearDown() method.

Tested that this fixes running ios_chrome_unittests with the
following command-line:

   --gtest_filter=MagicStackRankingModelTest.*:NewTabPageCoordinatorTest.*

Fixed: 441228775
Change-Id: Iadb1e6af42299e452f87cead49feb7cd40aa9753
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6884524
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507011}