Rename SelectActionMenuDelegate to TextSelectionCapabilitiesDelegate
SelectActionMenuDelegate is very similar to SelectionActionMenuDelegate
which is a separate interface that does quite different things. This
problem is exacerbated as the two types exist as parameters to the same
function. This CL renames SelectActionMenuDelegate with a more
descriptive name based on what it actually does which is return whether
to enable each of the default text selection capabilities.

Change-Id: Icfae46be3e0e61c3a69b930ae6548e05f1641334
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6884607
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Alex Mitra <alexmitra@chromium.org>
Auto-Submit: Alex Mitra <alexmitra@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507018}