Roll libc++ from a5fc4c8ab688 to 9f4b5a50d88f (1 revision)
This rolls in a change that changes behavior of std::multimap's and
std::multimap's find() method. find() used to return the first
found element for a key. Now it just returns any element with that
key. Use lower_bound() to get the first, or equal_range() to get
all elements for a given key.

We tried to find and update all places that (incorrectly) depended
on the old behavior, but it's of course possible we missed some.

See also chromium-dev thread
"Heads-up: std::multimap::find(), std::multiset::find() are changing behavior":
https://groups.google.com/a/chromium.org/g/chromium-dev/c/pPEi67j02L4

https://chromium.googlesource.com/external/github.com/llvm/llvm-project/libcxx.git/+log/a5fc4c8ab688..9f4b5a50d88f

2025-08-15 nikolasklauser@berlin.de [libc++] Optimize __tree::find and __tree::__erase_unique (#152370)

Bug: 441317299
Change-Id: I406ae2c19976e06d4aa7e4de0da4deeabeb50e59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6852893
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507035}