Temporarily disable the login-event test for proto-based reporting
See crbug.com/441339044 and crrev.com/c/6874508 for issues run into when
importing the relevant protos to Android and possible solutions.

Since this is the only blocker for the proto-based reporting launch and
we already test coverage for the log-in event[1], I decided to disable
the test for now to unblock the launch.

This CL also removes the reference to the already launched
`EnterpriseSecurityEventReportingOnAndroid` experiment.

[1]:https://source.chromium.org/chromium/chromium/src/+/main:components/enterprise/connectors/core/reporting_test_utils.cc;bpv=1;bpt=1;l=322?gsn=ExpectLoginEvent&gs=KYTHE%3A%2F%2Fkythe%3A%2F%2Fchromium.googlesource.com%2Fcodesearch%2Fchromium%2Fsrc%2F%2Fmain%3Flang%3Dc%252B%252B%3Fpath%3Dcomponents%2Fenterprise%2Fconnectors%2Fcore%2Freporting_test_utils.cc%23xNchOtWFCE1UDeWqHMvtPB45ErgVkPjpcyKIxojofQo

Bug: 441339044, 394403142
Change-Id: I2bfa809290b5512fc9aac4bf4bb0ed5e1ebb873f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6884807
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Commit-Queue: Nancy Xiao <nancylanxiao@google.com>
Reviewed-by: Shanthanu Bhardwaj <xanth@google.com>
Cr-Commit-Position: refs/heads/main@{#1507052}