Use SyncToken ref for GetBackingClientSharedImageForExternalWrite
Previously, GetBackingClientSharedImageForExternalWrite() takes a
SyncToken* out-parameter, and only generate new SyncToken if this
pointer is non-empty. Now that we always generate the new SyncToken in
GetBackingClientSharedImageForExternalWrite() (see the change in
https://crrev.com/c/6874649), it is no longer necessary to pass in a
pointer. So in this CL, we refactored the pointer into the more robust
SyncToken& type.

This CL also changes the position of the SyncToken& out-param to conform
to the Chromium style guide, and removes the unused
GetBackingClientSharedImageForOverwrite() function.

Bug: 40286368
Change-Id: I088e3704f6d9f3d5e519667d9fd4888aa7b4b5ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6879184
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507071}