[webui browser] Remove unneeded export, workarounds from SidePanelUIBase
In a previous CL, when we added SidePanelUIBase, missing method
definition linker errors caused us to mark the class with an export
macro. In subsequent CLs, however, this led to more subtle import/
export linker errors on Windows component builds. Workarounds using
NOINLINE and trivial pass-throughs seemed to fix the errors, but then
the errors would reappear in some form in a CL further along the
dependency chain.

After removing the initial export macro and subsequent workarounds,
it seems that the missing method errors were transient (and since
they were local I do not have a record of them on the bots). Hence
I think it may be safe to remove this code.

Bug: 439850831,439850993,393144165
Change-Id: Ib6d872185f046514daf321a7467ebabe72106412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6884811
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507109}