performance_manager: Boost priority of closing tabs
When a user closes a tab, its renderer process might be running at a low
priority if the tab was in the background. This can lead to a noticeable
delay in running unload handlers, making the UI feel unresponsive.

This change introduces `ClosingPageVoter`, a new priority voter that
observes the `IsClosing` state on `PageNode`s. When a page begins
closing, this voter casts a `USER_BLOCKING` priority vote for its main
frame. This ensures the renderer has the necessary resources to shut
down quickly and smoothly.

This behavior is controlled by the `kBoostClosingTabs` feature flag,
which is disabled by default.

Bug: 432275395

Change-Id: I9c6d1fb6de07fd90f9ea190f39d15bddeb3fe8ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6854025
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507137}