TaskAttribution: Instrument FileReader events
This CL adds task attrubtion support for FileReader events, capturing
the context when a read* method is called, and restoring it when events
are dispatched.

This also adds [[nodiscard]] to the new task_attribution_util helpers,
which bit me (forgetting to use the return value) while drafting this.

Fixed: 428017530
Change-Id: Ic06e0fcfc492e4f38886938d3c4fb77f1060a365
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6887311
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507187}