Extended expiration of compression dictionary histograms
This extends the expiration for the shared dictionary histograms until
the summer of 2026 (and assigns ownership of them to me).

We are still actively working on optimizing the dictionary fetch path,
dictionary storage and working through a rollout for enterprises
(where middleboxes don't interact well) and the histograms are needed
for that work.

Bug: 433257937,434693100,434692564,418772427,433257937,434693100,431772329,434692564,418772427
Change-Id: I421f302527fed8ac54a9fb265e7dcec3ab956bb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6885849
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1507221}